Run Directory for Artifacts (#80)
Added a run directory for storing artifacts (e.g. checkpoints, non-rank-0 logs, torch profiler traces)

Unlike most hparams, the CLI launch script needs to be aware of the run directory, so it can create output files for the non-rank-0 processes. However, it would result in a confusing api to have this parameter specified both via the launch script and via a trainer hparam.

Following the convention of other launch script arguments, the run directory is set as a CLI argument for the launch script, which then passes it through as a process-local environment variable.

Updated other components to use the run directory. Specifically:

Updated the wandb logger to automatically log files placed in the run directory as artifacts
Updated the checkpointer to store checkpoints inside the run directory by default
Updated the pytorch profiler to store traces in the run directory by default, so they are visible in wandb.