Timer PR parts 1, 2, and 7 from #146
Added the `Time` class, `Timer` class, coversion function, and test cases and docstrings for these classes.

Modified the `StringEnum` class to relax the requirement that the lowercase name be the value, since this does not hold for our timing abbreviations.

Other PRs will integrate the Timer into the training loop.