Timer PR parts 1 and 2 from #146 (#174)
Added the `Time` class, `Timer` class, conversion function, and test cases and docstrings for these classes. This functionality encompasses parts 1 and 2 from #146, with the following changes:

1. `tokens` will be 0 (instead of None) if not being tracked.

Also, modified the `StringEnum` class to relax the requirement that the lowercase name be the value, since this does not hold for our timing abbreviations.

Other PRs will integrate the Timer into the training loop.