Fix broken tests (#257)
Removed test_mnist_real_dataset since it's slow (downloads the entire mnist) and doesn't add much value
Refactored _convert_time_fields in the scheduler so it doesn't modify the dataclass in place. The issue with in-place modification was that checkpoint saver was saving the hparams dataclass after the fields have been converted. Then, when resuming from checkpoint, the already-converted fields were no longer valid time strings, so it crashed.
Added requests to the base dependency (it is used by the checkpoint loader)