Fixed all uses of textwrap.dedent
Textwrap.dedent requires each line to have the same common prefix (including the first line!). As such, each use of `textwrap.dedent` should be begin with `"""\` followed by a newline. For example:

```python
     textwrap.dedent("""\
         first line
         second line""")

This may have been causing bugs in jenkins with excessively long lines and/or newlines that could not be parsed properly.