Guard `dist.barrier` in the checkpointer with try/finally (#334)
If one process hits a distributed barrier but another one crashes before reaching it, then the process that hit the barrier will hang indefinitely. Distributed barriers do not respond to sigterm. Instead, this PR guards the code in the checkpointer that could raise errors with try/finally, so any exceptions (such as from failed downloads) will be re-propegated AFTER the barrier is released.

Closes #290