Remove requirement for `validation` and `metrics` (#378)
Users who just want to run training, and not validate on an external dataset should not need to supply validation and metrics as part of their `ComposerModel`. Especially now that the `eval_dataloader` is an optional argument to our trainer. This PR removes the abstract marker for the two methods in question.

Co-authored-by: Austin <A-Jacobson@users.noreply.github.com>