Skip saving of direct serialization fields
When the trainer is constructed, the passes in a max_duration, grad_accum, and precision. These values should take precedence over whatever was in a checkpoint, so there is no need to serialize them.

Since these were the only direct serialization attributes, also cleaned up the state logic to just require that the attributes to be serialized are listed. (Renamed `fields` to `attrs` to reflect that State is no longer a dataclass, since fields are a dataclass concept).

Also cleaned up the state serialization by refactoring the deepspeed logic out of the state and into the trainer next to where deepspeed is initialized. The state should not need to know whether the model is deepspeed or not.

Closes #441.