[Checkpointing - PR 5] Move the `CheckpointSaver` to a callback. (#687)
This PR implements the checkpoint saving functionality as a callback. See the [design doc](https://www.notion.so/Checkpoint-API-Redesign-f0ee88d4bd5149b6a091589ed3f465f5).

When reviewing this PR, please cross-reference the design document to ensure that all features are included.
Checkpoint metadata (#451) will be added in a later PR.

I realize this PR is kinda large, but the biggest diff comes from moving the `CheckpointSaver` to its own file, as a callback.