Set the random seed before each test.
Motivation behind using a deterministic seed for each test is to avoid test flakiness and ensure that PRs do not fail due to unrelated test failures. Instead, we should configure Jenkins to run nightly regressions across multiple seeds to identify tests that depend on the random initialization, and triage as appropriate.

1. Added a pytest CLI option `--seed`. If not set, defaults to 0.
2. Added an `autouse` fixture to set the seed automatically before each test.
3. For tests that create instances of the trainer, manually specified the `rank_zero_seed`, as otherwise the trainer would be re-seeded.
4. Removed the xfail on `test_masked_accuracy`, as the seed is now deterministic.
5. Unrelated cleanup: Removed unused fixtures from `conftest`; filtering warnings that should be ignored.

Closes #767