Remove the LoggerData and LoggerDataDict types
This PR switches the types of `LoggerData` to `Any` and `LoggerDataDict` to `Dict[str, Any]`.

While this change makes typing looser, it adds flexability for what users can log via custom logger callbacks. For example, with this change, a user could log an image, and also write a custom logger destination that knows how to handle images.