Serialize model state without `module.` prefix when using DDP (#829)
This PR strips the `module.` prefix from the checkpoint state when saving DDP checkpoints. Before, the `module.` prefix was consumed when loading checkpoints.

This change should improve usability, as `state_dict['state']['model']` reflects the original model that was passed to the trainer. The user should not have to worry about DDP wrapping.