Bind the schedulers to the state in `__init__()`, rather than on `fit()`.
Preferable to keep variables on the state object rather than as trainer members, where appropriate. Before, the state.schedulers was empty after `__init__()` but before `fit()`. Now, state.schedulers contains the compiled composer schedulers or original pytorch schedulers.

Restored optimizers on Event.INIT; it will be a bigger issue to rewrite algs to not depend on optimizers on init.