Add Computed Metrics to State (#842)
Add the current metrics to the state.
Reconfigured test case tolerances and deterministic mode, as the metrics are stored on the state. While deterministic mode works well for CPU tests, it does not fully work on GPU tests, especially for large models. So, the state test cases use tolerances for all tensor type comparisons.
Closes #830