Refactored console logging into the ProgressBarLogger
Console logging and the progress bar logger need to live together, so prints can be routed through `tqdm.write` as not to interfere with the progress bar.

As discussed offline, the progress bar logger parameters are flattened into `Trainer.__init__`, so it is possible to always append the progress bar logger (unless if it is explicitely set to False).