Console log level (#900)
Added back console logging. Before, this functionality lived in the FileLogger, which could previously write to `stdout`. That functionality was removed in #710, which introduced stdout and stderr capture to the file logger. This PR brings it back and merges it with the progress bar logger. Doing so in the same class allows prints to go through `tqdm.write`, so log lines do not interfere with the progress bar.

In addition, the progress bar logger parameters are flattened into `Trainer.__init__`. This change makes it is possible to always append the progress bar logger (unless if it is explicitly set to `False` via the `progress_bar: bool` argument).