Fix the launcher script
Killing by process group ID was accidentally killing the launcher script before it could terminate child processes. Instead, the launcher script will attempt the SIGTERM kill with just the PID. If that fails, then it will
attempt a SIGKILL kill with the PID and all child process PIDs. This is necessarry since, during a SIGKILL, the killed process cannot intercept the signal to manually terminate its children.

This approach also ensures that when the launcher script is killed via SIGKILL by the process group ID, all training processes also are killed.

Manually tested this approach in the terminal.