Require that optimizers are specified on `Trainer.__init__()`
While the original design called for being able to specify optimziers on `fit()`, that will not work with DeepSpeed or algorithms that perform surgery. (When not using DeepSpeed, the user would need to remember to construct the optimizer with `state.model.parameters()`, not `original_model.parameters()`, and that feels messy). As such, the Trainer will continue to
require that optimizers are specified when constructing the trainer. Happy to revisit this decision at a later point if there is a use case for changing the optimizers on each call to fit.