Multi Eval Improvements + `Trainer.eval()` (#947)
These changes help clean up the trainer code by moving the evaluator parameters into the evaluator. The implementation for multiple calls for fit and eval is based on this cleanup. These changes also allow one to do a full evaluation less frequently than a evaluation on a subset of the dataset.
    
Features:
- Added `subset_num_batches` to individual evaluators, so each evaluator can go through a different amount of the data
- Added `eval_interval` to individual evaluators, so evaluators can be called at different intervals
- Adjusted hparams to use `eval_interval`
- `Trainer.eval` can be used for eval only

Closes #329
Closes #40