Only pushy dockerimages if the branch is `dev`
Docker changes should only be pushed if the branch is `dev`, not `main`, as not all dockerfile changes may be cherry-picked in a maintance release. Since all commits already go on the dev branch, the docker images are already up-to-date, so they should not be updated again when the `main` branch is updated for releases.