Remove the Scale Schedule Algorithm (#854)
The algorithm form of scale schedule has been deprecated. It is available as an argument to the trainer.

This PR removes the Scale Schedule Algorithm as an algorithm (it must be specified via the trainer init args).
It also restores the scale schedule method card that was in the 0.3.1 release. This method card has been updated to
reflect the non-algorithm-class usage.

Closes #434.