W & B Checkpointing Example
When writing the release notes, I wrote an example for how to save and restore checkpoints from the W & B Logger. This example doesn't have any test coverage (need to figure that out later -- we don't have any test coverage for the `examples/` folder), but I did run it locally, and it works.