Move the model and optimizers to the device before `Event.INIT`
Current, the model and optimizers are moved to the device at the end when constructing the trainer. This change will allow callbacks and algorithms to read the device during `Event.INIT`, It will also help ensure that, when using non-cuda hardware, any errors with surgery will be raised when the algorithm is applied rather than later in the training loop.

- Updated the training loop
- Updated module surgery
- Added a test cases to ensure that the model is on the correct device