Add GitHub.com links in the API Reference
It's more useful to link the docs to the underlying source code rather than to a copy of it on the docs site. This PR replaces `sphinx-viewcode` with `sphinx-linkcode`, which adds links to each docstring entry to the source file and line number.

For module-level attributes, line numbers are determined by parsing the ast tree. For functions, classes, and function methods, `inspect.getsourcelines` is used to determine the line number of the definition.

`git` is used to determine the commit hash of the current build, so all docs links will be permalinks and work for both development and release branches.