AutoYAHP Part 4: Remove the Hparams! (#1043)
This PR removes Hparam classes for algorithms, callbacks, loggers, devices, and the profiler. (It leaves the hparam classes for models and datasets. It may be possible to remove the model hparam classes, but that can be done later).

Unlike before, hparam classes are no longer re-exported. This change will eventually allow us to remove `yahp` as a required dependency of Composer, without having to do the conditional imports everywhere.