Treat `DeprecationWarning` as a warning in tests, not error (#1119)
Deprecation or other warnings can occur whenever a new package version is released that we haven't tested against yet (even minor releases). For contributors, that means that their local pytest returns an error on parts of the codebase that is not part of their actual contribution, making it difficult to make progress.

e.g. someone contributing an algo would be forced to fix warnings in esoteric parts of the codebase. Or, end up figuring out how to add filterwarnings to an unrelated test, and make that judgement call.