Add devcontainer config (#1142)
Add a devcontainer config for development in a container when using vscode. Makes it easier to develop on the docker image, which has system level dependencies installed.
Using the mosaicml user to avoid permission issues that would occur when developing as root. However, sudo is still available if a user needs to switch into root inside the container.
Using the mosaicml/pytorch_vision:latest image, so it will work on gpu machines and so it can run all tests (including vision ones).