Require the `run_name` when `autoresume=True`
There is a nasty circular dependency when trying to make the run name optional with `autoresume=True`:

* Event.INIT needs the run name to be set.
* The run name could be in a cloud checkpoint, which would be fetched from the loggers.
* The Loggers are not fully initialized until Event.INIT runs

While it would be possible to autogenerate the run name AFTER Event.INIT if using autoresume, this would likely not help much in practice, as checkpoints would probably be stored in the cloud, where the run_name is being used. So, this PR requires that the `run_name` is always specified when `autoresume=True`.

Coming along for the ride: Some typing fixes