Fix the Composer Launch Script for the Composer Dockerimage
In the composer dockerimage, since the `composer` package is pip installed, the examples -- including the yaml entrypoint -- is not included. To fix, this PR moves the composer entrypoint into `composer.trainer.trainer_hparams` and  binds it to the `composer_train` CLI command.

So that `composer_train` can be used via the composer launcher script, this PR also adds in "command mode", where the user can pass an arbitrary executable that should be launched by the launcher script. For example:

```shell
composer -c composer_train -f composer/yamls/models/classify_mnist_cpu.yaml --train_dataset.mnist.datadir ~/datasets --val_dataset.mnist.datadir ~/datasets
```

Coming along for the ride: A fix in the docstrings that broke `composer_train --help`. Argparse treats help strings as python format strings, so all `%` will need to be escaped before passing it to argparse. This is a yahp bug which will be fixed later, but for now, just fixing the one offending docstring.