Always build+push Composer dockerimages
This PR always builds and pushes composer dockerimages to the `mosaicml/composer_builds` repo. This will help with debugging, so a composer image can be tested on non-release branches.

TODO: Determine the latency that this adds. Potentially only do this on the merge commits (so it does not slow down PR checks)