Build+push Composer dockerimages to `mosaicml/composer_staging` (#1197)
This PR builds and pushes composer dockerimages to the `mosaicml/composer_staging` repo on each merge commit. This will help with debugging, so a composer image can be tested on non-release branches. Each commit triggers a build of the cpu and gpu image, and both will be pushed to this repo as the following:

`mosaicml/composer_staging:${commitHash[:7]}`
`mosaicml/composer_staging:${commitHash[:7]}_cpu`

Coming along for the ride: fix the memory limit