Fix logging of hparams to object stores (#1297)
Fix the logging of hparams to object stores in the `examples/run_composer_trainer.py`. It was likely that the file was not being flushed before uploading, causing an empty file to be uploaded. To fix, switched to using a TemporaryDirectory, and creating a file within the tmpdir that can be properly closed before uploading.

Example of the fix: https://wandb.ai/mosaic-ml/ravi-composer-examples/runs/uexuga9v/files/config.yaml

Closes https://mosaicml.atlassian.net/browse/CO-687