Retry SFTP on Size Mismatch
On ocasssion, the SFTP Object Store could raise a file size mismatch.
This PR fixes that by manually validating the file size, and raising an `ObjectStoreTransientError` if the remote file size differs from the local file size. This behavior replicates the `confirm=True` flag that paramiko provides, but instead raises an error message that we can retry on.

Closes https://mosaicml.atlassian.net/browse/CO-642