Fix a race condition in the object store logger
The Object Store Logger had a race condition in `post_close` where the workers could be shut down before all files were enqueued, due to the same flag being used multiple times. Fixed by adding a second flag to trigger shutdown of the enquing thread that is seperate from the flag used to shutdown the worker pool.

Did not add a new test case to verify the "fix", as it is impossible to deterministically trigger this race condition via the public API.

Closes https://mosaicml.atlassian.net/browse/CO-638