Tutorial on training without Local Storage
Added a tutorial on training without local storage.

In addition to running all cells (to make sure it works), please give this tutorial a close read and comment on what is unclear.

My goal is to do a walkthrough showing how artifact logging, remote checkpointing, and streaming datasets all tie together. My hope is that as we refactor streaming datasets and logging, we can simplify this tutorial into something that is more intuitive.

Closes https://mosaicml.atlassian.net/browse/CO-758