Metrics Refactor Part 2 (#1419)
This PR implements the second half of the metrics refactor (#1411) that updates the training loop and all the Composer models.

The training loop now uses the previously added state.train_metrics and state.eval_metrics from #1411 to perform all training and evaluation on the correct set of metrics.

The main changes with respect to models is that the models now have split up validate() into eval_forward() and update_metrics() methods, which run an evaluation forward pass and update the metrics with the outputs of the evaluation forward pass respectively. This is mainly to get rid of the double forward pass (#467).