Remove log level and should_log_artifact (#1603)
Removes the should_log_artifact which was unused as far as we know, and would require a user writing a custom filter function that takes in state and LogLevel to decide whether to log an artifact. Also, the WandBLogger did not support this argument. should_log_artifact was also the only use of LogLevel, so that has been removed everywhere.