Object store logger refactor (#1601)
Renames everything in ObjectStoreLogger to (mostly) not use "object" and not use "artifact" at all, and instead refer to uploading/downloading files. The commit messages are clear on the "root" of each change (e.g. renaming the artifact_name argument to remote_file_name, and then each change touches many other files as a result