Usability improvements to `intitialize_dist()` (#1619)
Minor usability improvements to utils.dist.initialize_dist(). Moves `_get_device()` and `is_tpu_installed()` out of `trainer.py` and into a new file `utils.devices.py` It also allows `initialize_dist()` to accept a string or Device (instead of only a Device) and a float timeout (with a default value of 300.0 seconds) instead of a datetime.delta object with no default.