Fix many failing notebook tests (#1646)
Fixes a bunch of failing notebook tests, specifically:
huggingface_models.ipynb: un xfails this test, it passes for me locally
custom_speedup_methods.ipynb: Fixes usage of %%time (has to be the first line of a cell)
exporting_for_inference.ipynb: The export for inference callback did not properly move all input types to cpu, and so was failing. Also adds a new gpu unittest that would've caught the issue.
auto_grad_accum.ipynb: Skip this on cpu. It fails when I run it via pytest with a CUDA error, but works when I run it as a notebook. So not sure whether it will pass in CI or not.
migrate_from_ptl.ipynb: Added some required functions for eval/metrics
streaming_dataloader_facesynthetics.ipynb: Pin version of mmcv and mmcv-full
TPU_Training_with_composer.ipynb: skip because it needs tpu
training_without_local_storage.ipynb: Lots of fixes related to recent checkpointing changes. Ran locally to ensure it works, but not entirely sure it will work in CI.