Build Streaming CIFAR10 Factory Function (#1729)
This PR adds a `build_streaming_cifar10_dataloader()` factory function as part of the deyahpification effort. It also updates all the `build_*_cifar10_dataloader()` functions to return `DataSpec`s, which seems to be the convention now. `build_streaming_cifar10_dataloader()` defaults to version 1 of streaming, because we do not seem to have a version 2-compatible copy of CIFAR10.