Add logic for shifting labels before computing metrics (#1913)
Adds a `shift_labels` init argument to `HuggingFaceModel` class. This instructs the model whether to shift labels by one token before computing metrics, which mimics the way HF Causal LM classes handle labels when computing loss. This fixes the current implementation, which never does this shifting and produces incorrect metric results for Causal LMs.

If `shift_labels` is not specified, `HuggingFaceModel` will try to infer the correct behavior based on whether the model is an instance of a registered HF Causal LM class (or a subclass of one).