remove problematic lzma package from docker image
The version of lzma provided by the lzma package has a bug: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=700681
where it refuses to decompress if the file extension does not match,
causing the test to fail. This could be worked around but the `xz-utils`
package already provides a version of `lzma` that does not have this
issue, so that can be used instead.